# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Eyekontact API",
    "description": "Eyekontact API",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/blogImages/{file}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/uploads/images/{file}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/uploads/videos/{file}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/uploads/icons/{file}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}